requirements.txt
asgiref==3.8.1
cuid2==2.0.1
Django==5.1.1
django-debug-toolbar==4.4.6
hashids==1.3.1
python-dotenv==1.0.1
sqlparse==0.5.1